// packages/soporte-it/src/incidentes/dto/create-incidente.dto.d.ts
import { UrgenciaIncidente } from '../../entities/incidente.entity';
export declare class CreateIncidenteDto {
    equipoId: string;
    descripcion: string;
    urgencia?: UrgenciaIncidente;
    fechaReporte?: string;
    aplicacionesAfectadas?: string;
    accionesPrevias?: string;
}
//# sourceMappingURL=create-incidente.dto.d.ts.map